Started to make db observable
